separate object to isolate "cloudflare:workers"
