+++ packages/durable-event-iterator/src/index.ts
@@ -1,6 +1,5 @@
 export * from './builder'
 export * from './event-iterator'
-export * from './object'
 export * from './plugin'
 export * from './schemas'
 export * from './upgrade'
