+++ packages/durable-event-iterator/src/object/index.ts
@@ -1,0 +1,2 @@
+export * from './object'
+export * from './router'
