+++ packages/durable-event-iterator/src/object/router.ts
@@ -1,0 +1,18 @@
+import { os } from '@orpc/server'
+import {
+  experimental_HibernationEventIterator as HibernationEventIterator,
+} from '@orpc/server/hibernation'
+import {
+  experimental_HIBERNATION_EVENT_ITERATOR_ID_KEY as HIBERNATION_EVENT_ITERATOR_ID_KEY,
+} from '../consts'
+
+const base = os.$context<{ ws: WebSocket, ctx: DurableObjectState }>()
+
+export const experimental_durableEventIteratorObjectRouter = {
+  subscribe: base.handler(({ context }) => {
+    return new HibernationEventIterator<any>((id) => {
+      const attachment = context.ws.deserializeAttachment()
+      context.ws.serializeAttachment({ ...attachment, [HIBERNATION_EVENT_ITERATOR_ID_KEY]: id })
+    })
+  }),
+}
